main_program/requirements.txt
keras==2.7.0
matplotlib==3.4.3
mtcnn==0.1.1
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
scipy==1.7.2
streamlit==1.2.0
tensorflow==2.7.0
